Move the cut-finder settings up into  module.
