+++ circuit_knitting/cutting/__init__.py
@@ -38,6 +38,23 @@ Circuit Cutting
     instructions.CutWire
     instructions.Move
 
+Automatic Cut Finding
+~~~~~~~~~~~~~~~~~~~~~
+
+.. autosummary::
+    :toctree: ../stubs/
+    :nosignatures:
+
+    find_cuts
+
+.. autosummary::
+    :toctree: ../stubs/
+    :nosignatures:
+    :template: autosummary/class_no_inherited_members.rst
+
+    find_cuts.OptimizationParameters
+    find_cuts.DeviceConstraints
+
 Quasi-Probability Decomposition (QPD)
 =====================================
 
@@ -59,23 +76,6 @@ Quasi-Probability Decomposition (QPD)
     qpd.generate_qpd_weights
     qpd.decompose_qpd_instructions
     qpd.qpdbasis_from_instruction
-
-Automatic Cut Finding
-~~~~~~~~~~~~~~~~~~~~~
-
-.. autosummary::
-    :toctree: ../stubs/
-    :nosignatures:
-
-    find_cuts
-
-.. autosummary::
-    :toctree: ../stubs/
-    :nosignatures:
-    :template: autosummary/class_no_inherited_members.rst
-
-    cut_finding.OptimizationParameters
-    cut_finding.DeviceConstraints
 
 CutQC
 =====
@@ -99,12 +99,9 @@ from .cutting_decomposition import (
     cut_gates,
     PartitionedCuttingProblem,
 )
-from .find_cuts import find_cuts
 from .cutting_experiments import generate_cutting_experiments
 from .cutting_reconstruction import reconstruct_expectation_values
 from .wire_cutting_transforms import cut_wires, expand_observables
-from .cut_finding.quantum_device_constraints import DeviceConstraints
-from .cut_finding.optimization_settings import OptimizationParameters
 
 __all__ = [
     "find_cuts",
@@ -116,6 +113,4 @@ __all__ = [
     "PartitionedCuttingProblem",
     "cut_wires",
     "expand_observables",
-    "DeviceConstraints",
-    "OptimizationParameters",
 ]
