+++ circuit_knitting/cutting/cut_finding/__init__.py
@@ -8,8 +8,3 @@
 # that they have been altered from the originals.
 
 """Main automated cut finding functionality."""
-
-from .optimization_settings import OptimizationParameters
-from .quantum_device_constraints import DeviceConstraints
-
-__all__ = ["DeviceConstraints", "OptimizationParameters"]
