+++ circuit_knitting/cutting/cut_finding/cut_optimization.py
@@ -13,6 +13,8 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import numpy as np
 from dataclasses import dataclass
 from typing import cast
@@ -28,7 +30,8 @@ from .search_space_generator import (
 from .disjoint_subcircuits_state import DisjointSubcircuitsState
 from .circuit_interface import SimpleGateList, GateSpec
 from .optimization_settings import OptimizationSettings
-from .quantum_device_constraints import DeviceConstraints
+if TYPE_CHECKING:
+    from ..find_cuts import DeviceConstraints
 
 
 @dataclass
