+++ test/cutting/cut_finding/test_best_first_search.py
@@ -11,9 +11,7 @@ from circuit_knitting.cutting.cut_finding.cut_optimization import CutOptimizatio
 from circuit_knitting.cutting.cut_finding.optimization_settings import (
     OptimizationSettings,
 )
-from circuit_knitting.cutting.cut_finding.quantum_device_constraints import (
-    DeviceConstraints,
-)
+from circuit_knitting.cutting.find_cuts import DeviceConstraints
 from circuit_knitting.cutting.cut_finding.disjoint_subcircuits_state import (
     get_actions_list,
 )
